requirements.txt
Django==1.11
beautifulsoup4==4.6.0
requests
lxml
mecab-python3
pandas